refactor(rollout): move the driver-side async engines to rollout/manager/
They hold no model and cannot generate, so calling them engines made 'engine'
mean three things at once: the BaseRolloutEngine ABC, a concrete backend, and
these. They are the layer that owns admission, acceptance and disposal over
time — a manager.

  engine/asynchronous.py    -> manager/{buffers,batch,agentic}.py
  AsyncBatchRolloutEngine   -> BatchManager
  AsyncAgenticRolloutEngine -> AgenticManager

'manager' rather than 'scheduler' because this tree already calls three
unrelated things a scheduler: the LR scheduler (utils/scheduler_utils.py), the
diffusion noise scheduler, and SGLang's own scheduler subprocesses.

Adds manager/protocol.py recording the consumer surface both implementations
share, and splits the policy-free mechanisms (VersionedBuffer, PendingGroups,
root_of) out of the file that held both engines and all their machinery.
launch_ceiling stays with BatchManager, whose two trainers are its only callers,
so its ownership note remains accurate.

Pure move + rename; no behaviour change. This is groundwork for LIN-693, which
turns AgenticManager into a rank-0 role owning the rollout task queue.

Test plan: an AST comparison against the pre-move file proves all eight
top-level definitions byte-identical modulo the renames. compileall,
ruff check + format, check-recipe-targets (2434 paths) and
check-experimental-boundaries all pass.
