+++ unirl/rollout/engine/__init__.py
@@ -1,12 +1,11 @@
 """Rollout engines over the canonical ``Sample`` request type.
 
-Two halves of one design: ``synchronous.py`` records the worker-side sync contracts
-(``BaseRolloutEngine`` — the broad ABC including coordinator engines — and
-``SyncRolloutEngine``, the ``Sample`` → ``Sample`` refinement the per-backend
-subpackages implement); ``asynchronous.py`` records the driver side (the
-batch/agentic async engines and their mechanisms).
+``synchronous.py`` records the worker-side contracts (``BaseRolloutEngine`` — the
+broad ABC including coordinator engines — and ``SyncRolloutEngine``, the ``Sample``
+→ ``Sample`` refinement the per-backend subpackages implement). The driver side
+lives in ``../manager/``, which owns admission, acceptance and disposal over time
+and holds no model.
 
-Deliberately empty otherwise: importing the driver-side ``asynchronous`` module
-must stay ray/torch-free, so this init imports nothing and consumers import the
-halves directly.
+Deliberately empty otherwise: this init imports nothing, so consumers import the
+module they need directly and the manager package stays ray/torch-free.
 """
